requirements.txt
alembic==1.3.2
autopep8==1.4.4
awscli==1.16.310
Babel==2.8.0
blinker==1.4
boto3==1.10.46
botocore==1.13.46
Click==7.0
colorama==0.4.1
docutils==0.15.2
dominate==2.4.0
elasticsearch==7.1.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
pyasn1==0.4.8
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
PyYAML==5.2
rsa==3.4.2
s3transfer==0.2.1
six==1.13.0
SQLAlchemy==1.3.11
urllib3==1.25.7
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
